% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{dax.baseline}
\alias{dax.baseline}
\title{Title}
\usage{
dax.baseline(pastDax, currDate, pastObs = 1000)
}
\arguments{
\item{pastObs}{}
}
\value{
data,table object with predictions
}
\description{
Title
}
