% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daxshell}
\alias{daxshell}
\title{Title}
\usage{
daxshell(preds, currDate, weekday = "Wednesday", nameSeries = "DAX")
}
\arguments{
\item{nameSeries}{}
}
\value{
a data.table corresponding to the challenge's format requirements
}
\description{
Title
}
