% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{energy.baseline}
\alias{energy.baseline}
\title{Title}
\usage{
energy.baseline(
  currDate,
  hours = c(11, 15, 19),
  wdays = c(6, 7),
  pastObs = 100
)
}
\arguments{
\item{pastObs}{}
}
\value{
data,table object with predictions
}
\description{
Title
}
