% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{energyshell}
\alias{energyshell}
\title{Title}
\usage{
energyshell(
  preds,
  currDate,
  targetFullDate = FALSE,
  targetDiff = c(2, 3),
  weekday = "Wednesday",
  nameSeries = "energy"
)
}
\arguments{
\item{nameSeries}{}
}
\value{
a data.table corresponding to the challenge's format requirements
}
\description{
Title
}
