% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-rawdata.R
\name{format.dax}
\alias{format.dax}
\title{Converts to UTC}
\usage{
\method{format}{dax}(
  currDate,
  pathFrom = here("data", "dax", "raw"),
  pathTo = here("data", "dax")
)
}
\arguments{
\item{pathTo}{}
}
\description{
Converts to UTC
}
