% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-rawdata.R
\name{format.energy.byhour}
\alias{format.energy.byhour}
\title{Converts to UTC and aggregates by hour}
\usage{
\method{format}{energy.byhour}(
  currDate,
  pathFrom = here("data", "energy", "raw"),
  pathTo = here("data", "energy")
)
}
\arguments{
\item{pathTo}{}
}
\description{
Converts to UTC and aggregates by hour
}
