% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getquants}
\alias{getquants}
\title{Title}
\usage{
getquants(tser, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), na.rm = FALSE)
}
\arguments{
\item{tser}{Time Series}

\item{probs}{quantile levels}
}
\description{
Title
}
